+++ database/migrations/2022_01_12_163611_create_political_divisions_table.php
@@ -20,7 +20,7 @@ class CreatePoliticalDivisionsTable extends Migration
             $table->string('political_divisionc_id')->nullable();
 
             $table->string('name');
-            $table->tinyInteger('level');
+            $table->smallInteger('level');
             $table->tinyText('last_level');
